// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };

type PageDocumentDataSlicesSlice =
  | HeroSlice
  | ManifiestoSlice
  | ModalidadesSlice
  | ProcesoSlice
  | TerritorioSlice
  | DetalleEditorialSlice
  | ConfianzaSlice
  | ContactoSlice
  | RichTextSlice;

/**
 * Content for Page documents
 */
interface PageDocumentData {
  /**
   * Title field in *Page*
   */
  title: prismic.TitleField;

  /**
   * Slice Zone field in *Page*
   */
  slices: prismic.SliceZone<PageDocumentDataSlicesSlice> /**
   * Meta Title field in *Page*
   */;
  meta_title: prismic.KeyTextField;

  /**
   * Meta Description field in *Page*
   */
  meta_description: prismic.KeyTextField;

  /**
   * Meta Image field in *Page*
   */
  meta_image: prismic.ImageField<never>;
}

/**
 * Page document from Prismic
 */
export type PageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<Simplify<PageDocumentData>, "page", Lang>;

export type AllDocumentTypes = PageDocument;